Fix LengthUnit/isDefault test crash on Xcode 16
Fix for #451 issue.

This seems to be a bug in Xcode 16. See swift forum discussion for more details https://forums.swift.org/t/fatal-error-internal-inconsistency-no-test-reporter-for-test-case-argumentids/75666.

The test crashes because there are two copies of the same value `LengthUnit.em(.infinity)/LengthUnit.default`. Moved the check of LengthUnit.default into a separate test `LengthUnitTests/defaultValue` to resolve this crash.